# API/padelapp-api-model/src/main/resources/swagger/apiPadel2.yaml
openapi: 3.0.3
info:
  title: API Padelhg
  description: |
    This is a sample API for managing paddle court reservations, based on the OpenAPI 3.0 specification. This API allows performing operations related to paddle court reservations, such as creating, modifying, and canceling reservations.

  contact:
    email: ccoutomanul@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080
...
paths:
  /booking:
    post:
      tags:
        - booking
      summary: Add a new booking
      description: Add a new booking
      operationId: addBooking
      requestBody:
        description: Create a new booking
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputBookingDTO'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "405":
          description: Invalid input
...
  /booking/findByDate:
    get:
      tags:
        - booking
      summary: Finds bookings by date
      description: Finds bookings that fall between two given dates
      operationId: findBookingsByDates
      parameters:
        - name: startDate
          in: query
          description: Date from which the filter is applied
          required: true
          style: form
          explode: true
          schema:
            type: string
            format: date-time
            example: 2023-07-21T17:30:00Z
        - name: endDate
          in: query
          description: Date from which the filter is applied
          required: true
          style: form
          explode: true
          schema:
            type: string
            format: date-time
            example: 2023-07-22T17:30:00Z
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid date value
        "401":
          description: Unauthorized Error
        "405":
          description: Validation exception
  /booking/findByType:
    get:
      tags:
        - booking
      summary: Finds booking by type
      description: "Multiple tags can be provided with comma separated strings. Use\
        \ tag1, tag2, tag3 for testing."
      operationId: findBookingByType
      parameters:
        - name: bookingType
          in: query
          description: Tags to filter by
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/BookingType'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid tag value
  /booking/{bookingId}:
    get:
      tags:
        - booking
      summary: Find booking by ID
      description: Returns a single booking
      operationId: getBookingById
      parameters:
        - name: bookingId
          in: path
          description: ID of booking to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid ID supplied
        "401":
          description: Unauthorized operation
        "404":
          description: Booking not found
    put:
      tags:
        - booking
      summary: Update an existing booking
      description: Update an existing booking by id
      operationId: updateBooking
      parameters:
        - name: bookingId
          in: path
          description: ID of the booking to update
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated booking information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputBookingDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDTO'
        "400":
          description: Invalid booking ID supplied
        "401":
          description: Unauthorized
        "404":
          description: Booking not found
        "405":
          description: Validation exception
    delete:
      tags:
        - booking
      summary: Delete a booking by ID
      description: Delete an existing booking by ID
      operationId: deleteBooking
      parameters:
        - name: bookingId
          in: path
          description: ID of the booking to delete
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Booking deleted successfully
        "400":
          description: Invalid booking ID supplied
        "401":
          description: Unauthorized
        "404":
          description: Booking not found
  /resource/{id}:
    get:
      tags:
        - resource
      summary: Get resource by ID
      description: Returns a single resource by ID
      operationId: getResourceById
      parameters:
        - name: id
          in: path
          description: ID of the resource to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceDTO'
        "400":
          description: Invalid ID supplied
        "401":
          description: Unauthorized operation
        "404":
          description: Resource not found
...
  /festive/{idfestive}:
    get:
      security:
        - bearerAuth: []
      tags:
        - festive
      summary: Get festive by id
      description: ''
      operationId: getFestivebyId
      parameters:
        - name: idfestive
          in: path
          description: festive id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FestiveDTO'
        '400':
          description: Invalid festive id supplied
        '404':
          description: Festive not found
  /festive/{date}:
    get:
      tags:
        - festive
      summary: Get festive by id
      description: ''
      operationId: getFestivebyDate
      parameters:
        - name: date
          in: path
          description: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FestiveDTO'
        '400':
          description: Invalid festive id supplied
        '404':
          description: Festive not found
...
components:
  schemas:
...
    PlayerPosition:
      type: string
      description: player's favorite position
      enum:
        - DRIVE
        - REVERSE
        - INDIFFERENT
    ResourceDTOType:
      type: string
      description: booking type
      enum:
        - PADEL
        - ARENA
        - GYM
        - ECOWASH
    BookingType:
      type: string
      description: booking type
      enum:
        - TOURNAMENT
        - TRAINING
        - GAME
        - MATCH-TELEGRAM
        - LEAGUE
    BookingVisibility:
      type: string
      description: determines whether players can be joined or not
      enum:
        - PRIVATE
        - PUBLIC
    ResourceDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        number:
          type: integer
          format: int32
          example: 12
        name:
          type: string
          example: Pista central
        resourceDTOType:
          $ref: '#/components/schemas/ResourceDTOType'
        timeSlot:
          type: integer
          description: reservable time slot
          example: 90
        startTimeSlot:
          type: string
          format: time
          example: 09:00:00
        endTimeSlot:
          type: string
          format: time
          example: 00:00:00
        basePrice:
          type: number
          format: double
          example: 5.5
        daysInAdvance:
          type: integer
          format: int32
          example: 3
...
    UserBooking:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        email:
          type: string
          format: email
          example: john@email.com
        phone:
          type: string
          example: "12345"
        userPosition:
          $ref: '#/components/schemas/PlayerPosition'
...
    BookingDTO:
      type: object
      properties:
        idBooking:
          type: integer
          format: int64
          example: 10
        resourceDTO:
          $ref: '#/components/schemas/ResourceDTO'
        starDate:
          type: string
          format: date-time
          example: 2017-07-21T17:30:00Z
        endDate:
          type: string
          format: date-time
        ownerUser:
          $ref: '#/components/schemas/UserBooking'
        price:
          type: number
          format: double
        guessts:
          $ref: '#/components/schemas/UserBooking'
        bookingType:
          $ref: '#/components/schemas/BookingType'
        bookingVisibility:
          $ref: '#/components/schemas/BookingVisibility'
    InputBookingDTO:
      required:
        - endDate
        - resourceDTO
        - startDate
        - user
      type: object
      properties:
        resource:
          type: integer
          example: 6
        starDate:
          type: string
          format: date-time
          example: 2017-07-21T17:30:00Z
        endDate:
          type: string
          format: date-time
        user:
          type: integer
        bookingType:
          $ref: '#/components/schemas/BookingType'
        guessts:
          type: array
          items:
            type: string
            example: [juanca1223, manu34]
        bookingVisibility:
          $ref: '#/components/schemas/BookingVisibility'
        level:
          type: integer
          example: 5
    FestiveDTO:
      type: object
      properties:
        festiveId:
          type: integer
          format: int64
          example: 10
        date:
          type: string
          format: date-time
          example: 2023-03-28
        description:
          type: string
          example: Reconquista de Vigo
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header